fetch funfact from server
