+++ client/src/graphql/query/funFactQuery.js
@@ -2,8 +2,7 @@ import { gql} from '@apollo/client';
 export const FUN_FACT = gql`
     query FunFact {
         funFact {
-            title
-            description
+            fact
         }
     }
 `
